Allow reloading on SIGHUP
Many daemons are setup to reload on a HUP signal. Making use of
`tableflip` from Cloudflare, we can achieve the same. However, since
`tableflip` forks to a new process, but shares file descriptors, this
doesn't play too nicely with terminals. Thus, this is best suited to run
under systemd, which will track pacoloco across pids and correctly send
output to journald. One downside is that `tableflip` is not compatible
with Windows.