Add a purge strategy based on versions of package
Similar to the way `paccache` removes old versions of packages, we
implement a strategy to purge packages from the cache directory, keeping
at most the n most recent versions of a package.

To support this, we also modify the configuration format. The user now
declares a purge strategy (defaulted to the existing last access-time
based strategy) under the `purge_strategy` key. There is also a key
(`purge_keep_at_most`) to configure how many versions of a package to
keep in the cache. In addition, we allow the user to disable purging
entirely. The user declares what strategy they want employed by setting
the `purge_strategy` key to one of "time", "count", or "none".

I see three advantages to using this new strategy over managing it
externally with `paccache`. One is it doesn't require the user to create
and manage an external process of running `paccache`. Two is it will
work on systems where `paccache` cannot or will not be installed (e.g. a
non-Arch system where installing `paccache` is not easy). Three is it
removes any extra packages immediately after downloading a newer version
of a package, instead of waiting for a scheduled run of `paccache`.