Purge old files on startup
When starting up, if we're using the count purge strategy, we should go
through all the repos to determine if there are too many versions of a
package and remove the oldest ones. This is done synchronously at
startup, which might be the correct way to do this. Errors are logged,
but ignored, so even if it fails, pacoloco should still start up.