Handle '.sig' files when purging old files
Rarely, a package will have a detached signature file in the package
cache. If such a file exists we don't want to consider it when purging
old versions of a package, but we do want to remove it if it's matching
package file is to be removed. Thus, we do just that.